First non-working draft of pam_pwd role
This linux system role is meant to set a password policy for different
RHEL releases (7, 8 and 9 at the beginning). It should offer an
interface to implement the steps described in the solution articles
(login required):

 - Set a password policy in Red Hat Enterprise Linux 7
   URL: https://access.redhat.com/solutions/2808101

 - Set Password Policy/Complexity in Red Hat Enterprise Linux 8
   URL: https://access.redhat.com/solutions/5027331

It's not fully functional yet. I made some noob mistakes and have used
bad practise for sure. Since it is my very first attempt to create a
Linux System Role I need some help and guidance to improve my skill.
I publish this commit to gather feedback in an very early stage of the
role development. Any feedback is welcome and appreciated.

References:
  - Bug 2038620 - [RFE] [RHEL7/8] Create and provide RHEL System Role
    for managing a Password Policy
    URL: https://bugzilla.redhat.com/show_bug.cgi?id=2038620

  - Red Hat Enterprise Linux (RHEL) System Roles
    URL: https://access.redhat.com/articles/3050101

Signed-off-by: Joerg Kastning <joerg.kastning@uni-bielefeld.de>