Fixed syntax errors and conditionals
 - Fixed syntax error in command used by handler
 - Add some debug tasks to see values of registered variables
 - Changed conditionals and tested that the work as expected
   (at least on my machine)

At this point I use 'ignore_errors: yes', because I'm not sure what
would be a better setting for 'failed_when'. This is subject to
change.

I don't know how to add idempotency to the following task:

~~~
 - name: enable-feature with-faillock
   command: authselect enable-feature with-faillock
~~~

Obviously there are only enable-feature and disable-feature available
for authselect(8) but no way to check whether the feature is already
enabled or not. I have to investigate this.
See also: #1

Signed-off-by: Joerg Kastning <joerg.kastning@uni-bielefeld.de>